Fix gen_impl for low/high conversion

+++ code/vftintrml_l_s.h
@@ -1,3 +1,3 @@
 for (int i = 0; i < 2; i++) {
-  dst.dword[i] = (s32)a.fp32[i]; // rounding mode is not expressed in C
+  dst.dword[i] = (s64)a.fp32[i]; // rounding mode is not expressed in C
 }
